+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### 0.4.0
 * Fixed an issue where init command would generate an invalid yml for scripts.
+* Fixed an issue in misleading error message in v2 validation hook.
 
 ### 0.3.10
 * Added a BC check for incident fields - changing from version is not allowed.