+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### 0.4.0
 * Added a BC validation for changing incident field type.
 * Fixed an issue where init command would generate an invalid yml for scripts.
+* Fixed an issue in misleading error message in v2 validation hook.
 
 ### 0.3.10
 * Added a BC check for incident fields - changing from version is not allowed.