+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### 0.4.0
 * Added a validator for Indicator Types (reputation-*.json).
+* Added a BC validation for changing incident field type.
 * Fixed an issue where init command would generate an invalid yml for scripts.
 * Fixed an issue in misleading error message in v2 validation hook.