+++ CHANGELOG.md
@@ -8,11 +8,12 @@
 * Added a BC validation for changing incident field type.
 * Fixed an issue where init command would generate an invalid yml for scripts.
 * Fixed an issue in misleading error message in v2 validation hook.
+* Fixed an issue in v2 hook which now is set only on newly added scripts.
 
 ### 0.3.10
 * Added a BC check for incident fields - changing from version is not allowed.
 * Fixed an issue in create-content-artifacts where scripts in Packs in TestPlaybooks dir were copied with a wrong prefix.
-* Fixed an issue in v2 hook which now is set only on new added scripts
+
 
 ### 0.3.9
 * Added a validation that incident field can not be required.